requirements.txt
asyncio
fastapi
pydantic
pydantic_settings
uvicorn[standard]
python-multipart
google-cloud-aiplatform